ld.so: Add DT_GNU_OFFSET support
Note that the change to elf.h makes the loader binary-incompatible with
existing glibcs, so you will need to set LD_LIBRARY_PATH to point to a
libc.so.6 built by this branch. (Eventually the goal is to teach the
loader to do that automatically.)

Also add a little shell script to generate semistatic binaries.